requirements.txt
aiohttp           == 3.9.3
aioice            == 0.9.0
aiortc            == 1.7.0
aiosignal         == 1.3.1
async-timeout     == 4.0.3
attrs             == 23.2.0
av                == 11.0.0
cffi              == 1.16.0
cryptography      == 42.0.3
dnspython         == 2.6.0
frozenlist        == 1.4.1
google-crc32c     == 1.5.0
idna              == 3.6
ifaddr            == 0.2.0
multidict         == 6.0.5
pip               == 23.0.1
pycparser         == 2.21
pyee              == 11.1.0
pylibsrtp         == 0.10.0
pyOpenSSL         == 24.0.0
python-socketio   == 5.11.1
setuptools        == 65.5.0
typing_extensions == 4.9.0
yarl              == 1.9.4
